240916requirements.txt
absl-py==0.15.0
alembic==1.7.7
astunparse==1.6.3
attrs==22.2.0
autopage==0.5.2
bio==1.5.2
biopython==1.79
biothings-client==0.3.1
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
click==8.0.4
cliff==3.10.1
cmaes==0.9.0
cmd2==2.4.3
colorlog==6.8.2
cupy==9.6.0
cupy-cuda101==9.6.0
cycler==0.11.0
dataclasses==0.8
et-xmlfile==1.1.0
fair-esm==2.0.0
fastrlock==0.8.2
filelock==3.4.1
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.48.2
h5py==2.10.0
huggingface-hub==0.4.0
idna==3.6
imbalanced-learn==0.8.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.4.1
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mygene==3.2.2
nltk==3.6.7
numpy==1.18.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
optuna==3.0.6
packaging==21.3
pandas==1.1.5
pbr==6.0.0
Pillow==8.4.0
prettytable==2.5.0
protobuf==4.21.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-version==0.0.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.27.1
requests-oauthlib==1.4.0
rsa==4.9
s3transfer==0.5.2
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.2
six==1.15.0
SQLAlchemy==1.4.52
stevedore==3.5.2
tape-proteins==0.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.10.0
torchaudio==0.7.2
torchvision==0.8.2+cu101
tqdm==4.64.1
transformers==4.18.0
typing-extensions==3.7.4.3
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
